Remove AcadosDiffMpcLayer base class, rename acados_torch -> torch
Delete leap_c/layer.py (abstract AcadosDiffMpcLayer ABC). The setup logic
is inlined directly into AcadosDiffMpcLayerTorch.__init__.

Rename leap_c/acados_torch.py -> leap_c/torch.py.
AcadosDiffMpcLayerTorch now inherits only from torch.nn.Module, fixing the
MRO issue where super().__init__() hit nn.Module before the layer base.

Update all docs (8 references) and test imports (3 files) from
leap_c.acados_torch -> leap_c.torch.

Fix test_backward_nmpc to use existing diff_mpc fixtures instead of the
deleted diff_mpc_indefinite_hess fixtures.
