+++ pyproject.toml
@@ -5,54 +5,37 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "leap-c"
 version = "2025.0.0"
-dependencies = [
-  "acados-template",
-  "casadi",
-  "numpy>=1.26.0",
-]
+dependencies = ["acados-template", "casadi", "numpy>=1.26.0"]
 
 requires-python = ">=3.11"
 maintainers = [
-  { name = "Dirk Reinhardt", email = "dirk.p.reinhardt@ntnu.no" },
-  { name = "Jasper Hoffmann", email = "hoffmaja@informatik.uni-freiburg.de" },
-  { name = "Leonard Fichtner", email = "leonard.fichtner@imtek.uni-freiburg.de" },
+    { name = "Dirk Reinhardt", email = "dirk.p.reinhardt@ntnu.no" },
+    { name = "Jasper Hoffmann", email = "hoffmaja@informatik.uni-freiburg.de" },
+    { name = "Leonard Fichtner", email = "leonard.fichtner@imtek.uni-freiburg.de" },
 ]
 description = "Learning Predictive Control"
 readme = "README.md"
 license = { file = "LICENSE.txt" }
-keywords = [
-  "Model Predictive Control",
-]
+keywords = ["Model Predictive Control"]
 classifiers = [
-  "Development Status :: 4 - Beta",
-  "Programming Language :: Python",
+    "Development Status :: 4 - Beta",
+    "Programming Language :: Python",
 ]
 
 [project.optional-dependencies]
-test = [
-  "pytest",
-  "pytest-cov",
-]
+test = ["pytest", "pytest-cov"]
 torch = ["torch>=2.0.0"]
 jax = ["jax"]
-notebooks = [
-  "marimo>=0.13.0",
-]
+notebooks = ["marimo>=0.13.0"]
 docs = [
-  "sphinx",
-  "furo",
-  "myst-parser",
-  "sphinx-autoapi",
-  "sphinx-autobuild",
-  "sphinxcontrib-mermaid",
+    "sphinx",
+    "furo",
+    "myst-parser",
+    "sphinx-autoapi",
+    "sphinx-autobuild",
+    "sphinxcontrib-mermaid",
 ]
-dev = [
-  "leap_c[docs]",
-  "leap_c[test]",
-  "leap_c[torch]",
-  "pre-commit",
-  "ruff",
-]
+dev = ["leap_c[docs]", "leap_c[test]", "leap_c[torch]", "pre-commit", "ruff"]
 
 [tool.setuptools]
 packages = ["leap_c"]
@@ -77,20 +60,14 @@ convention = "google"
 
 [tool.ruff.lint]
 ignore = [
-  "E741",
-  "D100",
-  "D101",
-  "D102",
-  "D103",
-  "D104",
-  "D105",
-  "D106",
-  "D107",
+    "E741",
+    "D100",
+    "D101",
+    "D102",
+    "D103",
+    "D104",
+    "D105",
+    "D106",
+    "D107",
 ]
-extend-select = [
-    "D",
-    "I",
-    "NPY",
-    "UP045",
-    "E501",
-]
+extend-select = ["D", "I", "NPY", "UP045", "E501"]
